% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACT.R
\name{GE_ACT}
\alias{GE_ACT}
\title{Adaptive Combination Tests for Set-Based Gene-Environment Interactions}
\usage{
GE_ACT(G, Y, E, W, outcomeType, GeneID = NULL, maf_cut = 0,
  missing_cut = 0.15, nReps = 10000)
}
\arguments{
\item{G}{n by p matrix of p SNPs for n subjects}

\item{Y}{length n vector of outcome for n subjects}

\item{E}{length n vector of exposure for n subjects}

\item{W}{n by q data frame of q covariates to ajust for n subjects}

\item{outcomeType}{"B" for binary outcome and "C" for continuous outcome}

\item{GeneID}{length p vector indicating GeneID for each SNP, i.e which Gene does each SNP belong to? If you don't want a two-level analysis, you can ignore it}

\item{maf_cut}{SNPs with Minor Allele Frequency lower than this threshold 
would be removed from the set, by default 0, i.e no SNP removed}

\item{missing_cut}{SNPs with missing rate higher than this cut-off value
would be removed from the set, by default 0.15. Missing values would be imputed by average}

\item{nReps}{number of resampling to calculate ACT P Value, by default 10000. 
We recommend to use a small value, say 1000, to see the significance. If significant, then try larger value to get more accurate P Values.}
}
\value{
nGene: number of Genes

P.Values: A list of ACT P values
\item{p.ACT.Z.weight.pos}{ACT P value based on positive part of marginal Z statistics weighted by screening statistics}
\item{p.ACT.Z}{ACT P value based on marginal Z statistics weighted by screening statistics}
\item{p.ACT.Z.weight}{ACT P value based on marginal Z statistics}
\item{p.ACT.X}{ACT P value based on marginal Chi-Square statistics}
\item{p.ACT.X.weight}{ACT P value based on marginal Chi-Square statistics weighted by screening statistics}

Z.Statistics: Z Statistics used to calculate ACT P Values
\item{GE.interaction}{Z statistics of marginal GE interaction}
\item{GY.association}{Z statistics of marginal GY association}
\item{GE.correlation}{Z statistics of marginal GE correlation}
\item{Weight}{screening Z statistics used as weight,stronger of GY association/GE correlation}

GE.interaction,GY.association,GE.correlation summary 
\item{Estimate}{Coefficient Estimate}
\item{Std.Error}{Standard Error}
\item{z value}{Z statistics}
\item{Pr(>|z|)}{P Value}
}
\description{
This function returns P values of the interaction effect between a set of SNPs 
 and exposure status. P values include ACT, ACT weighted by screening statistics in both Z scale 
 and Chi-Square scale. Users need to input a set of genetic variants G, outcome Y, 
 exposure E, and a few covariates to asjust W and outcome type, "B" for binary and ="C" for continuous.
}
\details{
The input G, Y, E, W need to be numerical, with same number of observations.Note that Y and E cannot
contain missing values, and please remove those observations with missing in Y or E. 
We suggest users impute G before using this function, though SNPs with missing rate lower than 
missing_cut would be imputed in the function with mean value for that SNP.
}
\examples{
## Generate data for examples
set.seed(100)
G1=rbinom(2000,1,0.3)
G2=rbinom(2000,1,0.4)
G3=rbinom(2000,1,0.25)
G4=rbinom(2000,1,0.5)
G=cbind(G1,G2,G3,G4)
E=rbinom(2000,1,0.3)
Y=rbinom(2000,1,0.2)
GeneID=c("Gene1","Gene1","Gene2","Gene3")
## Example without covariates to adjust
GE_ACT(G,Y,E,W=NULL,outcomeType="B",GeneID)

## Example with covariates to adjust
age=rnorm(2000,45,2.6)
bmi=rnorm(2000,11.2,2.2)
GE_ACT(G,Y,E,W=cbind.data.frame(age,bmi),outcomeType="B")

## Example if users want to change default settings
Y=rnorm(2000,2,4)
GE_ACT(G,Y,E,W=cbind.data.frame(age,bmi),outcomeType="C",maf_cut=0.05,missing_cut=0.1)
}
\author{
Shirong Zhang  <shirongz@usc.edu>

Juan Pablo Lewinger  <lewinger@usc.edu>
}
\references{
Shirong Zhang, Juan Pablo Lewinger Adaptive set-based methods for Gene-Environment Interactions
}

