% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTGE.R
\name{GE_multiple}
\alias{GE_multiple}
\title{Multiple Testing for Gene-Environment Interactions (MTGE)}
\usage{
GE_multiple(G, Y, E, W, outcomeType, maf_cut = 0, missing_cut = 0.15)
}
\arguments{
\item{G}{n by p matrix of p SNPs for n subjects}

\item{Y}{length n vector of outcome for n subjects}

\item{E}{length n vector of exposure for n subjects}

\item{W}{n by q data frame of q covariates to ajust for n subjects}

\item{outcomeType}{"B" for binary outcome and "C" for continuous outcome}

\item{maf_cut}{SNPs with Minor Allele Frequency lower than this threshold 
would be removed from the set, by default 0, i.e no SNP removed}

\item{missing_cut}{SNPs with missing rate higher than this cut-off value
would be removed from the set, by default 0.15. Missing values would be imputed by average}
}
\value{
unadjusted and adjusted P values for each individual SNP
\item{p_unadjsted}{unadjusted P values}
\item{p_MTGE}{adjusted P values using MTGE}
\item{p_Sidak}{adjusted P values using Sidak}
}
\description{
This function returns P values of the interaction effect between individual SNPs 
 and exposure status. P values include unadjusted P values, P values adjusted by MTGE and Sidak. Users need to input a set of genetic variants G, outcome Y, 
 exposure E, and a few covariates to asjust W and outcome type, "B" for binary and ="C" for continuous.
}
\details{
The input G, Y, E, W need to be numerical, with same number of observations.Note that Y and E cannot
contain missing values, and please remove those observations with missing in Y or E. 
We suggest users impute G before using this function, though SNPs with missing rate lower than 
missing_cut would be imputed in the function with mean value for that SNP.
}
\examples{
## Generate data for examples
set.seed(100)
G1=rbinom(2000,1,0.3)
G2=rbinom(2000,1,0.4)
G3=rbinom(2000,1,0.25)
G4=rbinom(2000,1,0.5)
G=cbind(G1,G2,G3,G4)
E=rbinom(2000,1,0.3)
Y=rbinom(2000,1,0.2)

## Example without covariates to adjust
GE_multiple(G,Y,E,W=NULL,outcomeType="B")

## Example with covariates to adjust
age=rnorm(2000,45,2.6)
bmi=rnorm(2000,11.2,2.2)
GE_multiple(G,Y,E,W=cbind.data.frame(age,bmi),outcomeType="B")

## Example if users want to change default settings
Y=rnorm(2000,2,4)
GE_multiple(G,Y,E,W=cbind.data.frame(age,bmi),outcomeType="C",maf_cut=0.05,missing_cut=0.1)
}
\author{
Shirong Zhang  <shirongz@usc.edu>

Juan Pablo Lewinger  <lewinger@usc.edu>
}
\references{
Shirong Zhang, Juan Pablo Lewinger Multiple Testing for Gene-Environment Interactions
}

